Lighting and shadows and dust motes.
