+++ Assets/Images/Windows/Materials.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3e5687cb626e6b5458a764d66bb28b3e
+guid: 5cffca954db59e040b7bc4704c2781fd
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
