+++ Assets/Images/Windows/Materials/curtains.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 23fccf469d212dc4ba200164f0119024
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
