--- Assets/Scripts/CharacterCreation/CharacterPart.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c354d7f0eb057704fb840c4beaf4d61a
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
