+++ ProjectSettings/TagManager.asset
@@ -41,6 +41,3 @@ TagManager:
   - name: Default
     uniqueID: 0
     locked: 0
-  - name: Inside
-    uniqueID: 2866283963
-    locked: 0
