+++ UserSettings/Search.settings
@@ -7,9 +7,9 @@ onBoardingDoNotAskAgain = true
 showPackageIndexes = false
 showStatusBar = false
 scopes = {
-	"last_search.DA949528" = "ambient"
+	"last_search.DA949528" = "shadow dis"
 	"OpenInspectorPreview.DA949528" = "0"
-	"currentGroup.DA949528" = "scene"
+	"currentGroup.DA949528" = "all"
 }
 providers = {
 	adb = {
@@ -56,6 +56,7 @@ providers = {
 objectSelectors = {
 }
 recentSearches = [
+	"ha"
 ]
 searchItemFavorites = [
 ]
